+++ src/ui/mod.rs
@@ -7,7 +7,7 @@ use std::fmt::Write as _;
 use std::io::{Stderr, Write};
 use std::path::PathBuf;
 
-use crate::app_state::TereAppState;
+use crate::app_state::{CdResult, TereAppState};
 use crate::error::TereError;
 use crate::settings::{CaseSensitiveMode, GapSearchMode, SortMode};
 pub use action::{Action, ActionContext};
@@ -456,6 +456,30 @@ impl<'a> TereTui<'a> {
     fn change_dir(&mut self, path: &str) -> CTResult<bool> {
         //TODO: if there are no visible items, don't do anything?
         let res = match self.app_state.change_dir(path) {
+            Ok(res) => {
+                self.update_header()?;
+                match res {
+                    CdResult::Success => {
+                        // all good, clear info message
+                        self.info_message("")?;
+                        true
+                    }
+                    CdResult::MovedUpwards {
+                        target_abs_path: p,
+                        root_error: e,
+                    } => {
+                        // couldn't change to the expected folder because it was not reachable, but
+                        // we have changed to another (parent) folder, inform the user
+                        let p = p.display();
+                        if cfg!(debug_assertions) {
+                            self.error_message(&format!("Couldn't enter '{p}' ({e:?})"))?;
+                        } else {
+                            self.error_message(&format!("Couldn't enter '{p}' ({e})"))?;
+                        }
+                        false
+                    }
+                }
+            }
             Err(e) => {
                 if cfg!(debug_assertions) {
                     self.error_message(&format!("{e:?}"))?;
@@ -463,11 +487,6 @@ impl<'a> TereTui<'a> {
                     self.error_message(&format!("{e}"))?;
                 }
                 false
-            }
-            Ok(()) => {
-                self.update_header()?;
-                self.info_message("")?;
-                true
             }
         };
         self.redraw_main_window()?;
@@ -707,8 +726,12 @@ impl<'a> TereTui<'a> {
                             Action::ChangeSortMode => self.cycle_sort_mode()?,
 
                             Action::RefreshListing => {
-                                self.change_dir(".")?; //TODO: use 'current dir' instead of hardcoded '.' (?, see also pardir discussion elsewhere)
-                                self.info_message("Refreshed directory listing")?;
+                                //TODO: use 'current dir' instead of hardcoded '.' (?, see also pardir discussion elsewhere)
+                                if self.change_dir(".")? {
+                                    // only update info message if cd was successful, otherwise
+                                    // we're overwriting the error message
+                                    self.info_message("Refreshed directory listing")?;
+                                }
                             }
 
                             Action::Help => self.help_view_loop()?,